Never convert block brace to do.
Converting brace to do might change the semantic of a program. For example:

```
foo bar {
  1 + 2
}
```

Here the block is associated with `bar`. Converting it to `do/end` will associate it to `foo`.
Since it's a bit hard to determine when it's safe to do the conversion, for now we
don't do it anymore.