Don't halt on syntax errors when formatting multiple files.
Syntax errors can happen because of dummy test files or templates.
In any case syntax errors will be caught by a CI running Ruby tests,
it's not the job of the formatter to check valid Ruby code, just check
that files that can be formatted are formatted, or produce no change.