Add `rufo-mode.el` to README
I'm working on an additional emacs rufo plugin that uses RCS diffs to apply changes to the file inline. It also has first class bundler support. I think it's a reasonable alternative to the other emacs plugin already available, as it's a pretty different strategy. Let me know if you'd like to include it