format(Tests): introduce new test format
One major challenge working with the rufo codebase is difficulty in
reviewing and verifying the test cases. Because they are lots of
single line strings with `\n` characters, it's hard to tell what's
really going on.

This PR exists to break many of these tests out into the
`spec/source_specs` directory. Using a special magic syntax, we are
able to denote tests and options, and even name them. It's much easier
to look at a test and tell exactly what is inputted and expected as
the output.