Fix bug: rufo output does not satisfy rufo --check
This is actually something I ran into when applying this to our
codebase, where sometimes the output of rufo differs from the output of
rufo run multiple times.

Re-running the entire formatter until the result stops changing seems
very heavy handed, there must be a simpler fix for this. help?