Ignore vendor directory
Why: This is used for source code not maintained in the project that it is being run on so it does not make sense to include it. Also this is used commonly in CI tools for the installation of dependencies. Prior to this Rufo had to be configured to ignore this directory.